Added roles and permissions and laravel to typescript model generation

+++ app/Http/Controllers/PeertjesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Peertje;
 use App\Models\PeertjeLocation;
+use App\Models\User;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
@@ -11,6 +12,8 @@ class PeertjesController extends Controller
 {
 
     public function show() {
+        $this->authorize('viewAny', Peertje::class);
+
         $peertjes = Peertje::query()->whereHas('locations', function ($query) {
             $query->where('created_at', '>', now()->subDays(7));
         })->get();
@@ -21,6 +24,10 @@ class PeertjesController extends Controller
     }
 
     public function list() {
+        if(!\Auth::user()->can('manage-peertjes')) {
+            abort(403, 'Je hebt niet de rechten om peertjes te bewerken.');
+        }
+
         $peertjes = Peertje::all();
 
         return Inertia::render('Peertjes/List', [
@@ -29,6 +36,8 @@ class PeertjesController extends Controller
     }
 
     public function create(Request $request) {
+        $this->authorize('create', Peertje::class);
+
         $validated = $request->validate([
             'name' => 'required|string',
         ]);
@@ -41,11 +50,15 @@ class PeertjesController extends Controller
     }
 
     public function destroy(Request $request) {
+
         $validated = $request->validate([
             'id' => 'required|integer|exists:peertjes,id',
         ]);
 
         $peertje = Peertje::query()->findOrFail($validated['id']);
+
+        $this->authorize('delete', $peertje);
+
         $peertje->delete();
 
         return redirect()->route('peertjes.list');
